Basic UndoManager implemented.
Caveats: UndoManager works on the Workspace level. Not fully implemented, requires Workspace to be ICloneable and implement Clone().

As much as I would like Undo/Redo to only store the smallest amount of data possible with each change, I am currently thinking that storing a snapshot of the whole workspace is probably not going to be that memory intensive. We can come back to this if we run into issues.

Undo/Redo on the menu will crash the application, as the PropertyChanged event does not chain properly. On a Redo/Undo stack change, the MainWindowViewModel needs to notify properties have changed, but it's not directly involved in the changes. Therefore, it needs to listen to the WorkspaceViewModel and that listens to the UndoManager's events. Currently WorkspaceViewModel fires PropertyChanged events in the constructor, but this fails to help, since the events aren't listened to by the MainWindow until after it is constructed.

Also implemented a ServiceLocator, and scoped out some other services for other UI things. Will implement Dialog services and Stream services soon. This should almost allow us to save the workspace model and restore it. Need to investigate serializability on the model.