Added basic binary serialisation functionality to Workspace and Startblock.
Startblock - to send the program over the network.
Workspace - for save/load.

Binary serialisation is ideal while sending the program over the network. For saving and loading we could potentially use XML serialisation, but I was having trouble with it. I'll look into XML serialiation further on if we really want it in future.