Brett: please read the comments of this commit
NOTE: BRETT... as we discussed already, you seem to have replicated some code in VirtualMachine.cs  that I have already written in the executor class. Also you should also change the server to accept just the startblock, the workspace will contain alot of data that the server will not need.

The code in Virtual_Machine.RunProgram() would not currently work either, I did not intend to send the whole execution stack through the list parameter, It seems safer to pass in an auxiliary (empty) collection containing the blocks to perform after, and then add the contents of the returned collection to the execution stack... see the the executor class in the Model folder.

Also I'm not really seeing the issue with visibility that you were talking about. I switched the visibility of Virtual_Machine to internal (which is default) and it seems to compile fine. However if you still think we should have the classes public, we can adjust it.